refactor(cli): rename builtin cosign plugin package to sigstore
Align package naming with the sigstore handler module. Update all
imports and references in builtin.go and setup.go.
