+++ cli/internal/plugin/builtin/builtin.go
@@ -7,7 +7,7 @@ import (
 	filesystemv1alpha1 "ocm.software/open-component-model/bindings/go/configuration/filesystem/v1alpha1/spec"
 	helmdigest "ocm.software/open-component-model/bindings/go/helm/digest"
 	"ocm.software/open-component-model/bindings/go/plugin/manager"
-	"ocm.software/open-component-model/cli/internal/plugin/builtin/cosign"
+	"ocm.software/open-component-model/cli/internal/plugin/builtin/sigstore"
 	ocicredentialplugin "ocm.software/open-component-model/cli/internal/plugin/builtin/credentials/oci"
 	"ocm.software/open-component-model/cli/internal/plugin/builtin/input/dir"
 	"ocm.software/open-component-model/cli/internal/plugin/builtin/input/file"
@@ -54,8 +54,8 @@ func Register(manager *manager.PluginManager, filesystemConfig *filesystemv1alph
 	if err := rsa.Register(manager.SigningRegistry, filesystemConfig); err != nil {
 		return fmt.Errorf("could not register RSA signing plugin: %w", err)
 	}
-	if err := cosign.Register(manager.SigningRegistry); err != nil {
-		return fmt.Errorf("could not register Cosign signing plugin: %w", err)
+	if err := sigstore.Register(manager.SigningRegistry); err != nil {
+		return fmt.Errorf("could not register Sigstore signing plugin: %w", err)
 	}
 
 	return nil
