+++ cli/internal/plugin/builtin/sigstore/oidc_plugin_test.go
@@ -1,4 +1,4 @@
-package cosign
+package sigstore
 
 import (
 	"testing"
